--- .jshintrc
@@ -1,11 +1,0 @@
-{
-    "indent": 4,
-    "globalstrict": true,
-    "browser": true,
-    "globals": {
-        "require": false,
-        "module": true,
-        "exports": true,
-        "console": true
-    }
-}
