+++ .travis.yml
@@ -11,7 +11,6 @@ install:
   - pip install -r dev_requirements.txt --use-mirrors
   - pip install -q Django==$DJANGO --use-mirrors
   - pip install psycopg2==2.5.1
-  - npm install
 before_script:
   - sudo mkdir -p /var/log/vegphilly
   - sudo touch /var/log/vegphilly/gunicorn-general.log
@@ -31,6 +30,5 @@ before_script:
 script:
   - python -Wall manage.py test vegancity.PageLoadTest vegancity.FunctionalSearchTest || true
   - coverage run --source=vegancity manage.py test --exclude-integration-tests
-  - grunt jshint browserify:bundle
 after_success:
   - coveralls
