+++ ansible/roles/os/tasks/main.yml
@@ -14,25 +14,6 @@
 
 
 ######################################
-## nodejs
-## Source: https://github.com/azavea/ansible-nodejs
-######################################
-
-- name: Configure the NodeSource APT key
-  apt_key: url=https://deb.nodesource.com/gpgkey/nodesource.gpg.key state=present
-
-- name: Configure the NodeSource APT repositories
-  apt_repository: repo="deb https://deb.nodesource.com/node {{ ansible_distribution_release }} main"
-                  state=present
-
-- name: Install Node.js
-  apt: pkg=nodejs state=latest
-
-- name: Install Grunt
-  command: npm install --global grunt-cli creates=/usr/bin/grunt
-
-
-######################################
 ## supervisor
 ######################################
