--- package.json
@@ -1,20 +1,0 @@
-{
-  "name": "vegphilly",
-  "version": "0.0.1",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/vegphilly/vegphilly.com.git"
-  },
-  "dependencies": {
-    "backbone": "^1.1.2",
-    "lodash": "^2.4.1"
-  },
-  "devDependencies": {
-    "browserify": "3.2.0",
-    "grunt": "0.4.5",
-    "grunt-browserify": "1.3.2",
-    "grunt-cli": "^0.1.13",
-    "grunt-contrib-jshint": "^0.10.0",
-    "grunt-contrib-watch": "^0.6.1"
-  }
-}
