+++ templates/base.html
@@ -17,7 +17,6 @@
     <script type="text/javascript" src="//ajax.googleapis.com/ajax/libs/jquery/1.8.2/jquery.min.js"></script>
     <script src="//netdna.bootstrapcdn.com/twitter-bootstrap/2.2.1/js/bootstrap.min.js"></script>
 
-    <script type="text/javascript" src="{{ STATIC_URL }}js/bundle.js"></script>
     <script type="text/javascript">
       var googleAnalyticsTrackingId = '{{ GOOGLE_ANALYTICS_TRACKING_ID }}';
       if (googleAnalyticsTrackingId !== '') {
