--- vegancity/js/lib/jquery.js
@@ -1,2 +1,0 @@
-// Return global jQuery reference.
-module.exports = $;
