--- vegancity/js/src/profile_page.js
@@ -1,15 +1,0 @@
-'use strict';
-
-var $ = require('jquery');
-
-function init() {
-    $(document).ready(function() {
-      $(".gear").click(function(){
-        $(".profile-menu").toggleClass("open");
-      });
-    });
-}
-
-module.exports = {
-    init: init
-};
