+++ vegancity/static/js/profile_page.js
@@ -1,0 +1,5 @@
+$(document).ready(function() {
+  $(".gear").click(function(){
+    $(".profile-menu").toggleClass("open");
+  });
+});
