+++ vegancity/static/js/vendor_form.js
@@ -1,7 +1,4 @@
-'use strict';
-
-var $ = require('jquery');
-
+/*global $ */
 $(document).ready(function() {
 
     function sanitizePhoneNumber(number) {
