+++ vegancity/templates/vegancity/partials/vendors_sidebar.html
@@ -97,3 +97,8 @@
   <h4>Something missing?</h4>
   <a href="{% url 'new_vendor' %}" class="button">Add a Restaurant</a>
 </div>
+
+<script>
+</script>
+
+
