+++ vegancity/templates/vegancity/vendor_detail.html
@@ -142,8 +142,10 @@
 
 
 {% block body_extra %}
-<script type="text/javascript">
-(function() {
+  <script type="text/javascript" src="//maps.googleapis.com/maps/api/js?sensor=false"></script>
+  <script src="//cdnjs.cloudflare.com/ajax/libs/underscore.js/1.4.4/underscore-min.js"></script>
+  <script type="text/javascript" src="{{ STATIC_URL }}js/map.js"></script>
+  <script type="text/javascript">
     var vendors = [
         {
             id: "{{ vendor.id }}",
@@ -155,14 +157,11 @@
             longitude: "{{ vendor.location.x }}"
         }
     ];
-    var $ = require('jquery'),
-        map = require('vegancity/map'),
-        vendorMap = map.vendorMap;
     $(document).ready(function() {
         vendorMap.initialize("#map_canvas", vendors, "detail", true);
     });
-}());
-</script>
+
+  </script>
   <link href='http://fonts.googleapis.com/css?family=Cabin:400,700' rel='stylesheet' type='text/css'>
 <style>
